ref(hybridcloud): Move backlog metrics into their own task module
The two metrics tasks share no code path with delivery — only the model — so they
were sitting in deliver_webhooks.py purely because that is where the first draft
grew. Splitting them out leaves deliver_webhooks.py byte-identical to master and
gives the observation code its own module and test file.

Task names change with the module. Nothing has been deployed under the old names,
so there is no registered-name migration to worry about.

Refs INC-2420
